[OPE-81] Split RHIME scientific recipes
