+++ openghg_inversions/rhime/__init__.py
@@ -25,28 +25,28 @@ from openghg_inversions.observation_error import select_aggregation_error_mode
 
 from .builders import RhimeModelBuilder, RhimeModelBuilderContext, RhimeModelBuildResult
 from .materialization import materialize_pymc_inputs
-from .params import params_from_config, resolve_flux_sources
+from .params import params_from_config, resolve_flux_sources, resolve_rhime_options
 from .preparation import (
     assemble_rhime_inputs,
     build_rhime_basis,
     build_rhime_sensitivities,
     filter_rhime_observations,
     retrieve_or_reload_rhime_data,
-)
-from .runner import (
-    RhimeResult,
-    build_multisector_rhime_model,
-    build_standard_rhime_model,
-    make_multisector_rhime_result,
-    make_standard_rhime_result,
-    resolve_rhime_options,
-    run_rhime,
-    run_rhime_from_prepared_inputs,
-    run_rhime_multisector,
-    sample_rhime_model,
     with_prepared_rhime_sites,
 )
-from .sampling import RhimeSampler
+from .multisector import (
+    build_multisector_rhime_model,
+    make_multisector_rhime_result,
+    run_rhime_multisector,
+)
+from .outputs import RhimeResult
+from .prepared import run_rhime_from_prepared_inputs
+from .sampling import RhimeSampler, sample_rhime_model
+from .standard import (
+    build_standard_rhime_model,
+    make_standard_rhime_result,
+    run_rhime,
+)
 from .specs import RhimeOutputSpec, RhimeRunSpec
 
 __all__ = [
