+++ openghg_inversions/rhime/multisector.py
@@ -1,0 +1,260 @@
+"""Readable scientific recipe for a source-resolved multisector RHIME inversion."""
+
+from __future__ import annotations
+
+from pathlib import Path
+from typing import Any
+
+import arviz as az
+import xarray as xr
+
+from openghg_inversions._timing import log_timing, timer_seconds, timer_start
+from openghg_inversions.basis.basis_functions import BasisFunctions
+from openghg_inversions.inversion_data import RhimeMergedData, RhimePreparedInputs
+from openghg_inversions.models import (
+    RhimeLikelihoodBuilder,
+    RhimeModelSpec,
+    build_rhime_multisector_model_from_spec,
+)
+from openghg_inversions.models._rhime_flux import _select_sector_design
+
+from ._model_building import (
+    builtin_model_build_result,
+    validate_likelihood_builder,
+    validated_custom_model_build,
+)
+from .builders import (
+    RhimeModelBuilder,
+    RhimeModelBuilderContext,
+    RhimeModelBuildResult,
+    callable_metadata,
+    validate_model_build_result,
+)
+from .materialization import materialize_pymc_inputs
+from .outputs import RhimeResult, apply_output_bundle, make_multisector_output_bundle
+from .params import params_from_config, resolve_rhime_options
+from .preparation import (
+    assemble_rhime_inputs,
+    build_rhime_basis,
+    build_rhime_sensitivities,
+    filter_rhime_observations,
+    retrieve_or_reload_rhime_data,
+    with_prepared_rhime_sites,
+)
+from .sampling import RhimeSampler, sample_rhime_model
+from .specs import RhimeRunSpec
+
+
+def _validate_multisector_basis_layout(
+    basis_functions: BasisFunctions,
+    model_spec: RhimeModelSpec,
+    inv_inputs: xr.Dataset,
+) -> None:
+    """Require each retained source basis to match its prepared sector design."""
+    design = inv_inputs["H"]
+    region_layouts: list[tuple[str, str, int, int, bool]] = []
+    for sector in model_spec.sectors:
+        try:
+            source_basis = basis_functions.for_source(sector.flux_source)
+        except (KeyError, ValueError) as exc:
+            raise ValueError(
+                f"Sector {sector.name!r} requires source {sector.flux_source!r}, "
+                "but the retained basis has no matching source-specific basis."
+            ) from exc
+        state_dim = source_basis.operator.meta.state_dim
+        basis_index = source_basis.operator.basis_matrix.get_index(state_dim)
+        sector_design = _select_sector_design(
+            design,
+            sector=sector.name,
+            source=sector.flux_source,
+            variable_suffix=sector.variable_suffix,
+        )
+        prepared_state_dims = [str(dim) for dim in sector_design.dims if dim != "nmeasure"]
+        if len(prepared_state_dims) != 1:
+            raise ValueError(
+                f"Sector {sector.name!r} -> source {sector.flux_source!r} must have exactly "
+                f"one prepared state dimension; found {prepared_state_dims!r}."
+            )
+        prepared_index = sector_design.get_index(prepared_state_dims[0])
+        region_layouts.append(
+            (
+                sector.name,
+                sector.flux_source,
+                len(basis_index),
+                len(prepared_index),
+                basis_index.equals(prepared_index),
+            )
+        )
+    if any(not matches for _, _, _, _, matches in region_layouts):
+        details = ", ".join(
+            f"sector {sector!r} -> source {source!r}: basis has {basis_count} regions, "
+            f"prepared H has {prepared_count}" + ("" if matches else " with different coordinates")
+            for sector, source, basis_count, prepared_count, matches in region_layouts
+        )
+        raise ValueError(
+            "Retained source-specific basis state coordinates do not match prepared H; " + details + "."
+        )
+
+
+def build_multisector_rhime_model(
+    *,
+    prepared: RhimePreparedInputs,
+    model_inputs: xr.Dataset,
+    run_spec: RhimeRunSpec,
+    model_builder: RhimeModelBuilder | None = None,
+    likelihood_builder: RhimeLikelihoodBuilder | None = None,
+) -> RhimeModelBuildResult:
+    """Validate source-specific bases and build the concrete multisector graph."""
+    if model_builder is not None and likelihood_builder is not None:
+        raise ValueError("Pass either `model_builder` or `likelihood_builder`, not both.")
+    timing_start = timer_start()
+    _validate_multisector_basis_layout(prepared.basis_functions, run_spec.model, prepared.inv_inputs)
+    builder_context = RhimeModelBuilderContext(
+        prepared_inputs=prepared,
+        run_spec=run_spec,
+        multisector=True,
+    )
+    if model_builder is not None:
+        result = validated_custom_model_build(model_builder, context=builder_context)
+    else:
+        model = build_rhime_multisector_model_from_spec(
+            model_inputs,
+            run_spec.model,
+            **({} if likelihood_builder is None else {"likelihood_builder": likelihood_builder}),
+        )
+        result = builtin_model_build_result(model, model_spec=run_spec.model, multisector=True)
+    if likelihood_builder is not None:
+        validate_model_build_result(result, context=builder_context, builder_kind="likelihood")
+    log_timing("rhime.model_build", timer_seconds(timing_start), multisector=True)
+    return result
+
+
+def make_multisector_rhime_result(
+    *,
+    prepared: RhimePreparedInputs,
+    run_spec: RhimeRunSpec,
+    sampler: RhimeSampler,
+    model_build_result: RhimeModelBuildResult,
+    idata: az.InferenceData,
+    build_and_sample_seconds: float,
+    model_builder: RhimeModelBuilder | None = None,
+    likelihood_builder: RhimeLikelihoodBuilder | None = None,
+) -> RhimeResult:
+    """Construct a multisector result with its sector-aware output products."""
+    result = RhimeResult(
+        run_spec=run_spec,
+        model_spec=run_spec.model,
+        output_spec=run_spec.output,
+        inv_inputs=prepared.inv_inputs,
+        idata=idata,
+        sampler=sampler,
+        model=model_build_result.model,
+        basis_functions=prepared.basis_functions,
+        model_build_result=model_build_result,
+        output_metadata={"build_and_sample_seconds": build_and_sample_seconds},
+    )
+    builder_metadata = dict(model_build_result.metadata)
+    if model_builder is not None:
+        identity = callable_metadata(model_builder)
+        result.output_metadata["model_builder"] = identity
+        builder_metadata["model_builder"] = identity
+    if likelihood_builder is not None:
+        identity = callable_metadata(likelihood_builder)
+        result.output_metadata["likelihood_builder"] = identity
+        builder_metadata["likelihood_builder"] = identity
+
+    timing_start = timer_start()
+    output_bundle = make_multisector_output_bundle(
+        output_spec=run_spec.output,
+        run_spec=run_spec,
+        model_spec=run_spec.model,
+        idata=idata,
+        prepared=prepared,
+        country_file=run_spec.output.country_file,
+        variable_roles=model_build_result.variable_roles,
+        builder_metadata=builder_metadata,
+    )
+    log_timing(
+        "rhime.output_bundle_total",
+        timer_seconds(timing_start),
+        multisector=True,
+        output_format=run_spec.output.output_format,
+    )
+    apply_output_bundle(result, output_bundle)
+    return result
+
+
+def run_rhime_multisector(
+    *,
+    config_file: str | Path | None = None,
+    merged_data: RhimeMergedData | None = None,
+    likelihood_builder: RhimeLikelihoodBuilder | None = None,
+    **kwargs: Any,
+) -> RhimeResult:
+    """Run the source-resolved multisector RHIME scientific recipe.
+
+    The visible process is resolve → retrieve/reload → filter → basis →
+    sensitivities → assemble → materialize → build → sample → result/output.
+    This module keeps source layout validation and sector-aware outputs beside
+    that process instead of hiding them behind standard/multisector branching.
+    """
+    validate_likelihood_builder(likelihood_builder)
+    params = (
+        params_from_config(config_file, extra_kwargs=kwargs, normalise=False)
+        if config_file is not None
+        else dict(kwargs)
+    )
+    setup = resolve_rhime_options(params=params, multisector=True)
+
+    preparation_start = timer_start()
+    merged = retrieve_or_reload_rhime_data(
+        setup.data_args,
+        multisector=True,
+        **({} if merged_data is None else {"merged_data": merged_data}),
+    )
+    filtered = filter_rhime_observations(merged, setup.data_args)
+    basis_functions = build_rhime_basis(filtered, setup.data_args)
+    site_data = build_rhime_sensitivities(
+        filtered,
+        basis_functions,
+        setup.data_args,
+        multisector=True,
+    )
+    prepared = assemble_rhime_inputs(filtered, basis_functions, site_data, setup.data_args)
+    log_timing(
+        "rhime.prepare_inputs",
+        timer_seconds(preparation_start),
+        multisector=True,
+        nmeasure=prepared.inv_inputs.sizes.get("nmeasure"),
+        sites=len(prepared.sites),
+        regions=prepared.inv_inputs.sizes.get("region"),
+        sources=prepared.inv_inputs.sizes.get("source"),
+        basis_source=prepared.basis_artifact_source,
+    )
+    run_spec = with_prepared_rhime_sites(setup.run_spec, prepared)
+
+    model_inputs = materialize_pymc_inputs(
+        prepared,
+        aggregation_error_mode=run_spec.model.aggregation_error_mode,
+    )
+    build_and_sample_start = timer_start()
+    model_build_result = build_multisector_rhime_model(
+        prepared=prepared,
+        model_inputs=model_inputs,
+        run_spec=run_spec,
+        likelihood_builder=likelihood_builder,
+    )
+    idata = sample_rhime_model(
+        model_build_result,
+        setup.sampler,
+        use_variable_roles=likelihood_builder is not None,
+    )
+    return make_multisector_rhime_result(
+        prepared=prepared,
+        run_spec=run_spec,
+        sampler=setup.sampler,
+        model_build_result=model_build_result,
+        idata=idata,
+        build_and_sample_seconds=timer_seconds(build_and_sample_start),
+        likelihood_builder=likelihood_builder,
+    )
