+++ openghg_inversions/rhime/standard.py
@@ -1,0 +1,208 @@
+"""Readable scientific recipe for a standard single-sector RHIME inversion.
+
+The public recipe deliberately spells out its complete execution order. Small
+amounts of forwarding shared with the multisector recipe are retained so a
+scientist can read or copy this module without reconstructing a pipeline.
+"""
+
+from __future__ import annotations
+
+from pathlib import Path
+from typing import Any
+
+import arviz as az
+import xarray as xr
+
+from openghg_inversions._timing import log_timing, timer_seconds, timer_start
+from openghg_inversions.inversion_data import RhimeMergedData, RhimePreparedInputs
+from openghg_inversions.models import RhimeLikelihoodBuilder, build_rhime_model_from_spec
+
+from ._model_building import (
+    builtin_model_build_result,
+    validate_likelihood_builder,
+    validated_custom_model_build,
+)
+from .builders import (
+    RhimeModelBuilder,
+    RhimeModelBuilderContext,
+    RhimeModelBuildResult,
+    callable_metadata,
+    validate_model_build_result,
+)
+from .materialization import materialize_pymc_inputs
+from .outputs import RhimeResult, apply_output_bundle, make_standard_output_bundle
+from .params import params_from_config, resolve_rhime_options
+from .preparation import (
+    assemble_rhime_inputs,
+    build_rhime_basis,
+    build_rhime_sensitivities,
+    filter_rhime_observations,
+    retrieve_or_reload_rhime_data,
+    with_prepared_rhime_sites,
+)
+from .sampling import RhimeSampler, sample_rhime_model
+from .specs import RhimeRunSpec
+
+
+def build_standard_rhime_model(
+    *,
+    prepared: RhimePreparedInputs,
+    model_inputs: xr.Dataset,
+    run_spec: RhimeRunSpec,
+    model_builder: RhimeModelBuilder | None = None,
+    likelihood_builder: RhimeLikelihoodBuilder | None = None,
+) -> RhimeModelBuildResult:
+    """Build the concrete single-sector graph and describe its output roles."""
+    if model_builder is not None and likelihood_builder is not None:
+        raise ValueError("Pass either `model_builder` or `likelihood_builder`, not both.")
+    timing_start = timer_start()
+    builder_context = RhimeModelBuilderContext(
+        prepared_inputs=prepared,
+        run_spec=run_spec,
+        multisector=False,
+    )
+    if model_builder is not None:
+        result = validated_custom_model_build(model_builder, context=builder_context)
+    else:
+        model = build_rhime_model_from_spec(
+            model_inputs,
+            run_spec.model,
+            **({} if likelihood_builder is None else {"likelihood_builder": likelihood_builder}),
+        )
+        result = builtin_model_build_result(model, model_spec=run_spec.model, multisector=False)
+    if likelihood_builder is not None:
+        validate_model_build_result(result, context=builder_context, builder_kind="likelihood")
+    log_timing("rhime.model_build", timer_seconds(timing_start), multisector=False)
+    return result
+
+
+def make_standard_rhime_result(
+    *,
+    prepared: RhimePreparedInputs,
+    run_spec: RhimeRunSpec,
+    sampler: RhimeSampler,
+    model_build_result: RhimeModelBuildResult,
+    idata: az.InferenceData,
+    build_and_sample_seconds: float,
+    model_builder: RhimeModelBuilder | None = None,
+    likelihood_builder: RhimeLikelihoodBuilder | None = None,
+) -> RhimeResult:
+    """Construct a standard result and write only its requested products."""
+    result = RhimeResult(
+        run_spec=run_spec,
+        model_spec=run_spec.model,
+        output_spec=run_spec.output,
+        inv_inputs=prepared.inv_inputs,
+        idata=idata,
+        sampler=sampler,
+        model=model_build_result.model,
+        basis_functions=prepared.basis_functions,
+        model_build_result=model_build_result,
+        output_metadata={"build_and_sample_seconds": build_and_sample_seconds},
+    )
+    builder_metadata = dict(model_build_result.metadata)
+    if model_builder is not None:
+        identity = callable_metadata(model_builder)
+        result.output_metadata["model_builder"] = identity
+        builder_metadata["model_builder"] = identity
+    if likelihood_builder is not None:
+        identity = callable_metadata(likelihood_builder)
+        result.output_metadata["likelihood_builder"] = identity
+        builder_metadata["likelihood_builder"] = identity
+
+    timing_start = timer_start()
+    output_bundle = make_standard_output_bundle(
+        output_spec=run_spec.output,
+        run_spec=run_spec,
+        model_spec=run_spec.model,
+        idata=idata,
+        prepared=prepared,
+        country_file=run_spec.output.country_file,
+        sampler=sampler,
+        variable_roles=model_build_result.variable_roles,
+        builder_metadata=builder_metadata,
+    )
+    log_timing(
+        "rhime.output_bundle_total",
+        timer_seconds(timing_start),
+        multisector=False,
+        output_format=run_spec.output.output_format,
+    )
+    apply_output_bundle(result, output_bundle)
+    return result
+
+
+def run_rhime(
+    *,
+    config_file: str | Path | None = None,
+    merged_data: RhimeMergedData | None = None,
+    likelihood_builder: RhimeLikelihoodBuilder | None = None,
+    **kwargs: Any,
+) -> RhimeResult:
+    """Run the standard single-sector RHIME scientific recipe.
+
+    The visible process is resolve → retrieve/reload → filter → basis →
+    sensitivities → assemble → materialize → build → sample → result/output.
+    ``merged_data`` may supply a borrowed retrieval-stage handoff, while
+    ``likelihood_builder`` is a direct-Python model customization seam.
+    """
+    validate_likelihood_builder(likelihood_builder)
+    params = (
+        params_from_config(config_file, extra_kwargs=kwargs, normalise=False)
+        if config_file is not None
+        else dict(kwargs)
+    )
+    setup = resolve_rhime_options(params=params, multisector=False)
+
+    preparation_start = timer_start()
+    merged = retrieve_or_reload_rhime_data(
+        setup.data_args,
+        multisector=False,
+        **({} if merged_data is None else {"merged_data": merged_data}),
+    )
+    filtered = filter_rhime_observations(merged, setup.data_args)
+    basis_functions = build_rhime_basis(filtered, setup.data_args)
+    site_data = build_rhime_sensitivities(
+        filtered,
+        basis_functions,
+        setup.data_args,
+        multisector=False,
+    )
+    prepared = assemble_rhime_inputs(filtered, basis_functions, site_data, setup.data_args)
+    log_timing(
+        "rhime.prepare_inputs",
+        timer_seconds(preparation_start),
+        multisector=False,
+        nmeasure=prepared.inv_inputs.sizes.get("nmeasure"),
+        sites=len(prepared.sites),
+        regions=prepared.inv_inputs.sizes.get("region"),
+        sources=prepared.inv_inputs.sizes.get("source"),
+        basis_source=prepared.basis_artifact_source,
+    )
+    run_spec = with_prepared_rhime_sites(setup.run_spec, prepared)
+
+    model_inputs = materialize_pymc_inputs(
+        prepared,
+        aggregation_error_mode=run_spec.model.aggregation_error_mode,
+    )
+    build_and_sample_start = timer_start()
+    model_build_result = build_standard_rhime_model(
+        prepared=prepared,
+        model_inputs=model_inputs,
+        run_spec=run_spec,
+        likelihood_builder=likelihood_builder,
+    )
+    idata = sample_rhime_model(
+        model_build_result,
+        setup.sampler,
+        use_variable_roles=likelihood_builder is not None,
+    )
+    return make_standard_rhime_result(
+        prepared=prepared,
+        run_spec=run_spec,
+        sampler=setup.sampler,
+        model_build_result=model_build_result,
+        idata=idata,
+        build_and_sample_seconds=timer_seconds(build_and_sample_start),
+        likelihood_builder=likelihood_builder,
+    )
